merge two arrays into one and print
Write a function that will merge two arrays into a new one and print the elements of the new array. The function will take the two arrays and their sizes as arguments.

for example:
int a = {1,2,3,4};
int b = {13,23,33,34};

New array will be c = {1,2,3,4,13,23,33,34} after the function merges the above two arrays and makes a new array c.